requirements.txt
api-display-purposes==0.0.3
APScheduler==3.6.3
arabic-reshaper==2.1.3
asgiref==3.5.2
astroid==2.11.6
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
bootstrap-admin==0.4.4
build==0.8.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
clarifai==2.6.2
click==8.1.3
colorama==0.4.5
configparser==3.8.1
cryptography==37.0.2
decorator==4.4.2
dill==0.3.5.1
Django==3.2
django-admin-thumbnails==0.2.6
django-bootstrap4==22.1
django-ckeditor==6.4.2
django-countries==7.3.2
django-credit-cards==0.4.1
django-crispy-forms==1.11.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-mptt-admin==2.2.0
django-parler==2.3
django-recaptcha==3.0.0
django-rosetta==0.9.8
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
EasyProcess==1.1
emoji==1.7.0
ez-setup==0.9
future==0.18.2
googleapis-common-protos==1.56.1
googletrans==3.0.0
grpcio==1.46.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
huepy==1.2.1
hyperframe==5.2.0
idna==2.8
imageio==2.19.2
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instabot==0.117.0
isort==5.10.1
jsmin==2.2.2
jsonschema==2.6.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==1.0.3
numpy==1.22.4
packaging==21.3
pep517==0.12.0
Pillow==9.1.1
pip-tools==6.8.0
platformdirs==2.5.2
pluggy==1.0.0
plyer==2.0.0
polib==1.1.1
proglog==0.1.10
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
PySocks==1.7.1
pytest==7.1.2
python-bidi==0.4.2
python-decouple==3.4
python-gettext==4.0
python-telegram-bot==13.11
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.4.24
reportlab==3.6.1
requests==2.22.0
requests-toolbelt==0.9.1
responses==0.21.0
rfc3986==1.5.0
schedule==1.1.0
selenium==3.141.0
semantic-version==2.9.0
setuptools-rust==1.3.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
tqdm==4.64.0
translate-po==1.0.13
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.25.11
webdriverdownloader==1.1.0.3
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.5
django-ipware==4.0.2
ip2geotools==0.1.6
